Fix for gcc

+++ fittest.cc
@@ -59,7 +59,7 @@ void fitTrack(const Track & trk, const TrackExtra& trkextra, int itrack, Event& 
 #ifdef INWARDFIT
   for (int i = trkLayers.size()-1; i >= 0; i--){
 #else
-  for (int i = 0; i < trkLayers.size(); i++){
+  for (int i = 0; i < static_cast<int>(trkLayers.size()); i++){
 #endif
     //for each hit, propagate to hit radius and update track state with hit measurement
     const Hit& hit = evt_lay_hits[trkLayers[i]][trk.getHitIdx(trkLayers[i])];
